refactor: move useActiveTab hook to a separate file and clean up layout components
